backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==22.0.0
whitenoise==6.7.0
